requirements.txt
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
google-api-core==0.1.4
google-auth==1.4.1
google-cloud-core==0.28.0
google-cloud-speech==0.32.0
google-cloud-storage==1.7.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.9.1
idna==2.6
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydub==0.20.0
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
